+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Look at disabled 'set-state-in-effect' and 'exhaustive-deps', might have to do with 'value cannot be modified' or 'cannot access refs during render' lint errors
 - vitest browser tests: close timed out after 20000ms
 - test coverage is by test config file, not integrated
+- added click listeners on range-inputs to test them, remove vitest/browser supports change listeners on inputs 
 
 ### Features
 - Playmode Tough: Play without flags or mine counters